+++ tests/unit/_bootstrap.php
@@ -1,0 +1,2 @@
+<?php
+include __DIR__ . '/../../vendor/rhubarbphp/rhubarb/platform/execute-test.php';